==> Framework Update 2021/04_analysis_loo_oom.R <==
rf.loo.oom <- function (mtryval) {
  
  save.loo.rf.oom <- vector(mode="list", length=nrow(d1))
  save.loo.pred.oom <- vector(mode="list", length=nrow(d1))
  
  for (jj in 1:(nrow(d1))) {
    
    current.train <- filter(d1, ii != jj)
    current.test  <- filter(d1, ii == jj)
    
    current.rf <- randomForest(cluster.OOM.rev ~ Crystal_Structure_rev + Crystal_Type_rev +
                                 Length_rev + Structural_Form_rev + Scale_rev + PP_size_nm_rev +
                                 Contaminants_ + Contaminant_Type + Contaminant_Amount +
                                 Functionalized_Type + Purification_Type + Modification +
                                 Solubility + Zeta_Potential + Surface_Charge + Density + 
                                 Surface_Area + Median_Aerodynamic_Diameter + Diameter +
                                 Agglomerated_ + Material_Category + material,
                               data=current.train,
                               importance=T,
                               proximity=T,
                               mtry=mtryval)
    save.loo.rf.oom[[jj]] <- current.rf
    
    current.pred <- predict(current.rf, newdata=current.test)
    save.loo.pred.oom[[jj]] <- current.pred
    
  } # jj loop end
  
  #update file name for MTRY
  saveRDS(save.loo.rf.oom, file=paste0(pathout, paste0("save.loo.rf.oom.mtry",mtryval,".RDS")))
  saveRDS(save.loo.pred.oom, file=paste0(pathout, paste0("save.loo.pred.oom.mtry",mtryval,".RDS")))
  
} #function end

start_time <- Sys.time()
for (cc in 1:22) {
  rf.loo.oom(mtry=cc)
}
end_time <- Sys.time()
end_time - start_time #31 minutes
